+++ modules/randomwalk/ext/RandomWalksLib/SparseSolverEigenLLT.cpp
@@ -34,7 +34,7 @@ std::vector<double> SparseSolverEigenLLT::solve_Ax_b(SparseMatrix<double> A, Spa
 		xmat[uidx[i]] = val;
 	}
 
-	for (int i=0; i<seeds->size(); i++)
+	for (size_t i=0; i<seeds->size(); i++)
 	{
 		if((*labels)[i] == active_label)
 			xmat[(*seeds)[i]] = 1.0;
